src/python/user-group-automate/pyproject.toml
[tool.poetry]
name = "user_group_automate"
version = "1.2.0"
description = "An util to generate the "
authors = ["CSITEA OY  <info@flok.fi>"]
maintainers = ["CSITEA OY  <info@flok.fi>"]

packages = [
    { include = "console_utils/*" }
]


[tool.poetry.dependencies]
python = ">3.8,<4"
colorama = "^0.4.4"
numpy="^1.23.0"
pandas = "^1.4.2"
matplotlib = "^1"
pyyaml = "^5.3.1"


[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.4"
pyyaml = "^5.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
